pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "CLI tool to generate and orchestrate full-stack projects."
requires-python = ">=3.11"
dependencies = [
    "copier>=9.0.0",
    "questionary>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "tomlkit>=0.14.0",
    "libcst>=1.8.6",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pyyaml>=6.0",
    "ruff>=0.7.0",
    "ty>=0.0.1a7",
    "pre-commit>=3.7.0",
]

[project.scripts]
forge = "forge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["forge*"]

[tool.setuptools.package-data]
"forge.templates" = ["**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov --cov-report=term-missing"
markers = [
    "e2e: end-to-end tests that scaffold a project and run its native test suite (slow, opt-in via -m e2e)",
]

[tool.coverage.run]
source = ["forge"]
omit = [
    "forge/templates/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
fail_under = 75
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
include = ["forge"]
exclude = ["forge/templates", "**/__pycache__", "**/.venv"]

[tool.ty.rules]
unresolved-import = "error"
invalid-argument-type = "error"
invalid-return-type = "error"
missing-argument = "error"
unknown-argument = "error"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    "forge/templates",
    ".venv",
    "build",
    "dist",
    "htmlcov",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # function call in argument default — common in argparse helpers
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
